rasengan/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
redis==5.2.1
httpx==0.28.1
